Add assumptions about setup/run and ports
Expand on how we envision the two key stages in this application:

*Setup*

Is assumed to be able to bring up all services and applications in the
correct order in a healthy condition, with all deployment commands
executed during the process.

Specifically, it means:
- services are up
- migrations and other one-off tasks were run
- applications are up
- healthchecks are green

*Run*

The running of the tests does not assume anything about how applications
are deployed; it uses a set of default URLs. All of these URLs can be
overridden.

If all URLs point to external services, there is no need to run the
setup step.