Add feature for provider upload testing
As the page in question is using AngularJS I had to put in some extra
code that waits for the dynamic elements to render on the pages.

Since providers cannot re-upload the same file using the form fields at
the top of the page, I had to put a branch in the step definition:

- one branch if the provider has never uploaded that month before
- and one for one that has already been uploaded

This branching was required as I think we should make sure these tests
are executable against other environments, where we can't throw away the
database to always start from a fresh state. Additionally, there is no
"delete" statement so we can't dynamically clean up either.